Update bmalloc()
Created an excetion dealer and getting the fitting size.\nNeed to work on iterating through to find memory depending on our policy (Bestfit, Firstfit).\nAlso need a way to get the starting address of the first block. It's returned from mmap(), but don't know where to recieve the void* return type.